refactor(tests/utils): move create_file_and_commit to tests/utils
