+++ tests/commands/test_bump_command.py
@@ -1,20 +1,9 @@
 import sys
-import uuid
-from pathlib import Path
-from typing import Optional
 
 import pytest
 
 from commitizen import cli, cmd, git
-
-
-def create_file_and_commit(message: str, filename: Optional[str] = None):
-    if not filename:
-        filename = str(uuid.uuid4())
-
-    Path(f"./{filename}").touch()
-    cmd.run("git add .")
-    git.commit(message)
+from tests.utils import create_file_and_commit
 
 
 @pytest.mark.usefixtures("tmp_commitizen_project")
